Code reuse through hook
